+++ src/components/SlackAdvancedFilters.tsx
@@ -1,0 +1,108 @@
+/**
+ * Slack検索の詳細フィルター条件コンポーネント
+ * - チャンネル、投稿者、期間の詳細条件設定
+ * - 展開・折りたたみ機能
+ * - 各フィルター条件の入力フィールド
+ */
+
+'use client'
+
+type SlackAdvancedFiltersProps = {
+  showAdvanced: boolean
+  onToggleAdvanced: () => void
+  channel: string
+  onChannelChange: (channel: string) => void
+  author: string
+  onAuthorChange: (author: string) => void
+  startDate: string
+  onStartDateChange: (date: string) => void
+  endDate: string
+  onEndDateChange: (date: string) => void
+  disabled?: boolean
+}
+
+export function SlackAdvancedFilters({
+  showAdvanced,
+  onToggleAdvanced,
+  channel,
+  onChannelChange,
+  author,
+  onAuthorChange,
+  startDate,
+  onStartDateChange,
+  endDate,
+  onEndDateChange,
+  disabled = false
+}: SlackAdvancedFiltersProps) {
+  return (
+    <div className="space-y-4 pt-2">
+      <button
+        type="button"
+        onClick={onToggleAdvanced}
+        className="text-sm text-blue-600 hover:text-blue-800 focus:outline-none"
+      >
+        {showAdvanced ? '詳細な条件を閉じる ▲' : 'もっと詳細な条件を追加する ▼'}
+      </button>
+      {showAdvanced && (
+        <div className="space-y-4 p-4 border border-gray-300 rounded-md bg-gray-50">
+          <div>
+            <label htmlFor="channel" className="block text-sm font-medium text-gray-700 mb-1">
+              チャンネル (例: #general)
+            </label>
+            <input
+              id="channel"
+              type="text"
+              value={channel}
+              onChange={(e) => onChannelChange(e.target.value)}
+              placeholder="#general"
+              className="block w-full px-3 py-2 border border-gray-400 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-1 focus:ring-blue-400 focus:border-blue-400 disabled:bg-gray-100 disabled:cursor-not-allowed transition-colors"
+              disabled={disabled}
+            />
+          </div>
+          <div>
+            <label htmlFor="author" className="block text-sm font-medium text-gray-700 mb-1">
+              投稿者 (例: @user)
+            </label>
+            <input
+              id="author"
+              type="text"
+              value={author}
+              onChange={(e) => onAuthorChange(e.target.value)}
+              placeholder="@user"
+              className="block w-full px-3 py-2 border border-gray-400 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-1 focus:ring-blue-400 focus:border-blue-400 disabled:bg-gray-100 disabled:cursor-not-allowed transition-colors"
+              disabled={disabled}
+            />
+          </div>
+          <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
+            <div>
+              <label htmlFor="startDate" className="block text-sm font-medium text-gray-700 mb-1">
+                投稿期間 (開始日)
+              </label>
+              <input
+                id="startDate"
+                type="date"
+                value={startDate}
+                onChange={(e) => onStartDateChange(e.target.value)}
+                className="block w-full px-3 py-2 border border-gray-400 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-1 focus:ring-blue-400 focus:border-blue-400 disabled:bg-gray-100 disabled:cursor-not-allowed transition-colors"
+                disabled={disabled}
+              />
+            </div>
+            <div>
+              <label htmlFor="endDate" className="block text-sm font-medium text-gray-700 mb-1">
+                投稿期間 (終了日)
+              </label>
+              <input
+                id="endDate"
+                type="date"
+                value={endDate}
+                onChange={(e) => onEndDateChange(e.target.value)}
+                className="block w-full px-3 py-2 border border-gray-400 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-1 focus:ring-blue-400 focus:border-blue-400 disabled:bg-gray-100 disabled:cursor-not-allowed transition-colors"
+                disabled={disabled}
+              />
+            </div>
+          </div>
+        </div>
+      )}
+    </div>
+  )
+}
